+++ lib/utils/corner_type.dart
@@ -1,1 +1,0 @@
-enum Corner { topLeft, topRight, bottomLeft, bottomRight }