+++ CHANGELOG.md
@@ -1,3 +1,11 @@
-## 0.0.1
+## 0.0.1 - 2024-09-24
 
-* TODO: Describe initial release.
+* Uploaded package
+
+## [0.0.2] - 2024-09-24
+### Added
+- Added license, better decription.
+
+## [0.0.3] - 2024-09-26
+### Added
+- Added pushReplacement, pushAndRemoveUntil, updated decription.